Reset and change password
Here we implement a full password reset and change system without using a real email service, instead we use a email console to do this.
Unit test for each case when resetting and changing password  

NOTE: urlsafe_base64_encode now return a string instead of a bytestring so urlsafe_base64_encode(######).decode() is no necessary. This was a problem for me inside a test_ view_password_reset.py